mai-poc/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
openai>=1.0.0
pillow>=10.0.0
pydantic>=2.0.0
